Move i_kubernetes_service to spark job dir
