+++ piezo_web_app/PiezoWebApp/src/services/spark_job/spark_job_service.py
@@ -1,6 +1,6 @@
 from kubernetes.client.rest import ApiException
 
-from PiezoWebApp.src.services.kubernetes.i_kubernetes_service import IKubernetesService
+from PiezoWebApp.src.services.spark_job.i_kubernetes_service import IKubernetesService
 from PiezoWebApp.src.utils.return_status import StatusCodes
 
 # str | The custom resource's group name
