+++ piezo_web_app/PiezoWebApp/tests/handlers/base_handler_test.py
@@ -7,7 +7,7 @@ from tornado.httpclient import HTTPClientError
 from tornado.testing import AsyncHTTPTestCase
 from tornado.web import Application
 
-from PiezoWebApp.src.services.kubernetes.i_kubernetes_service import IKubernetesService
+from PiezoWebApp.src.services.spark_job.i_kubernetes_service import IKubernetesService
 from PiezoWebApp.src.utils.route_helper import format_route_specification
